Task#27 Creating database structure for question.
