+++ src/main/java/sda/ispeak/prework/Runner.java
@@ -5,17 +5,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 import sda.ispeak.prework.models.dtos.user.UserDto;
-import sda.ispeak.prework.models.questions.Answer;
-import sda.ispeak.prework.models.questions.Question;
-import sda.ispeak.prework.models.questions.QuestionSubject;
 import sda.ispeak.prework.models.users.User;
-import sda.ispeak.prework.repositories.AnswerRepository;
 import sda.ispeak.prework.repositories.QuestionRepository;
 import sda.ispeak.prework.repositories.UserRepository;
 import sda.ispeak.prework.services.UserService;
-
-import java.util.HashSet;
-import java.util.Set;
 
 @Component
 @Slf4j
@@ -24,9 +17,6 @@ public class Runner implements CommandLineRunner {
 
     @Autowired
     QuestionRepository questionRepository;
-
-    @Autowired
-    AnswerRepository answerRepository;
 
     @Autowired
     UserService userService;
@@ -52,25 +42,6 @@ public class Runner implements CommandLineRunner {
             userDto.setConfirmedPassword(user.getPassword());
 
             userService.save(userDto);
-
-            Answer answerA = new Answer("Good morning", true);
-            Answer answerB = new Answer("Guten Dag", false);
-
-            Answer save = answerRepository.save(answerA);
-            Answer answer = answerRepository.save(answerB);
-
-
-            Set<Answer> answers = new HashSet<>();
-            answers.add(save);
-            answers.add(answer);
-            Question question = new Question("PresentSimple", "Jak jest dzien dobry po angielsku?");
-
-
-            questionRepository.save(question);
-
-            Set<Answer> answers1 = question.getAnswers();
-
-            QuestionSubject futureSimple = QuestionSubject.valueOf("FutureSimple");
 
 
         } catch (Exception e) {
