--- src/main/java/sda/ispeak/prework/models/questions/QuestionSubject.java
@@ -1,16 +1,0 @@
-package sda.ispeak.prework.models.questions;
-
-public enum QuestionSubject {
-    PresentSimple,
-    PresentContinuous,
-    PresentPerfect,
-    PresentPerfectContinuous,
-    PastSimple,
-    PastContinuous,
-    PastPerfect,
-    PastPerfectContinuous,
-    FutureSimple,
-    FutureContinuous,
-    FuturePerfect,
-    FuturePerfectContinuous
-}
